Fix issues with the date the item is closed.
The date an item (pr/issue) was closed was not accurate, but instead was
based on when the item was most recently touched/changed.  Now the date
the issue/pr is closed/done is based on those time values.

Add more tests to the code as well.

Remove the draft item support since it will never be done in that form.